chore: merge BaseLauncherSettings into LauncherSettings
For historic reasons, the launcher settings classes were split into an abstract class interface and (several) implementations, but we dropped most of those some time ago.

Since we now only have single implementation class there's no point in keeping the interface and implementation separate.
